Test user balance ratio in relation to events that happen in the month

+++ app/src/main/java/com/chamwari/tech/xitique/domain/entities/BalanceSummary.kt
@@ -2,7 +2,7 @@ package com.chamwari.tech.xitique.domain.entities
 
 
 data class BalanceSummary(
-    val relativeBalanceInPercentage: Float,
+    val relativeBalanceInPercentage: Int,
     val balanceMessage: String,
     val balance : Int
 )
